pyproject.toml
[project]
name = "huggingface-paper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "lxml>=6.0.0",
    "python-dotenv>=1.1.1",
]
